US334402: extract all classes, strucs, enums and types into individual swift files
