+++ AccessCheckoutSDK/AccessCheckoutSDK/ApiHeaders.swift
@@ -1,0 +1,4 @@
+final class ApiHeaders {
+    public static let verifiedTokensHeaderValue = "application/vnd.worldpay.verified-tokens-v1.hal+json"
+    public static let sessionsHeaderValue = "application/vnd.worldpay.sessions-v1.hal+json"
+}
