+++ AccessCheckoutSDK/AccessCheckoutSDK/CVV.swift
@@ -1,0 +1,2 @@
+/// The card CVV
+public typealias CVV = String
