+++ AccessCheckoutSDK/AccessCheckoutSDK/CVVSessionURLRequestFactory.swift
@@ -15,24 +15,3 @@ class CVVSessionURLRequestFactory {
         return request
     }
 }
-
-private struct UserAgent {
-    static let headerName = "X-WP-SDK"
-    static let valueFormat = "access-checkout-ios/%@"
-    let headerValue: String
-
-    init(bundle: Bundle) {
-        let appVersion = bundle.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String ?? "unknown"
-        self.headerValue = String(format: UserAgent.valueFormat, appVersion)
-    }
-}
-
-struct SessionsSessionRequest: Codable {
-    enum Key: String, CodingKey {
-        case cvc = "cvc"
-        case identity = "identity"
-    }
-
-    var cvc: String
-    var identity: String
-}
