+++ AccessCheckoutSDK/AccessCheckoutSDK/CardTextView.swift
@@ -1,0 +1,6 @@
+/// A view representing a text field
+public protocol CardTextView: CardView {
+    
+    /// The text
+    var text: String? { get }
+}
