+++ AccessCheckoutSDK/AccessCheckoutSDK/ExpiryMonth.swift
@@ -1,0 +1,2 @@
+/// The card expiry date month
+public typealias ExpiryMonth = String
