+++ AccessCheckoutSDK/AccessCheckoutSDK/ValidationResult.swift
@@ -1,0 +1,9 @@
+/// Result of validation
+public struct ValidationResult {
+    
+    /// Result is partially valid
+    public let partial: Bool
+    
+    /// Result is completely valid
+    public let complete: Bool
+}
